api/userに関する型定義を一つにまとめる
