+++ frontend/src/lib/api/user.ts
@@ -2,11 +2,8 @@ import { AxiosPromise } from 'axios'
 
 import client from 'lib/api/client'
 
-import type { CreateUserParams } from 'types/createUserParams'
-import type { CreateUserResponse } from 'types/createUserResponse'
+import type { CreateUserParams, CreateUserResponse, UpdateUserParams, UpdateUserResponse } from 'types/api/user'
 import type { FirebaseIdToken } from 'types/firebaseIdToken'
-import type { UpdateUserParams } from 'types/updateUserParams'
-import type { UpdateUserResponse } from 'types/updateUserResponse'
 
 export const createUser = (params: CreateUserParams, idToken: FirebaseIdToken): AxiosPromise<CreateUserResponse> =>
   client.post('/users', params, {
