--- frontend/src/types/createUserParams.ts
@@ -1,5 +1,0 @@
-export type CreateUserParams = {
-  name: string | null
-  avatar?: { data: string; name: string }
-  invitationToken: string | null
-}
