--- frontend/src/types/createUserResponse.ts
@@ -1,7 +1,0 @@
-import type { User } from 'types/user'
-
-export type CreateUserResponse = {
-  user: User
-  invitationToken: string
-  isTeamCapacityReached: boolean
-}
