--- frontend/src/types/updateUserParams.ts
@@ -1,4 +1,0 @@
-export type UpdateUserParams = {
-  name: string
-  avatar?: { data: string; name: string }
-}
