--- frontend/src/types/updateUserResponse.ts
@@ -1,5 +1,0 @@
-import type { User } from 'types/user'
-
-export type UpdateUserResponse = {
-  user: User
-}
